core/vision_engine.py
import cv2
import numpy as np
from typing import Dict, Any, List
from pathlib import Path
from datetime import datetime
import json
from utils.logger import logger

class VisionEngine:
    """
    Core Vision Engine for JARVIS AI.
    Handles video analysis, object detection, and anomaly detection in visual data.
    """
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.vision_enabled = config.get("VISION_ENABLED", False)
        self.model_name = config.get("VISION_MODEL", "yolo-v8-nano")
        self.cache_size_mb = config.get("VIDEO_ANALYSIS_CACHE_SIZE_MB", 500)
        self.models_loaded = False
        self.recording_active = False # Placeholder for future live recording feature

        if self.vision_enabled:
            self._load_models()
            logger.info(f"Vision Engine initialized with model: {self.model_name}")
        else:
            logger.info("Vision Engine is disabled in configuration.")

    def _load_models(self):
        """
        Loads pre-trained computer vision models (e.g., YOLO for object detection).
        This is a placeholder for actual model loading.
        """
        try:
            # Simulate loading a model
            logger.info(f"Simulating loading vision model: {self.model_name}...")
            # In a real scenario, you'd load models using libraries like
            # TensorFlow, PyTorch, or OpenCV's DNN module.
            # Example: self.model = cv2.dnn.readNetFromDarknet(cfg_path, weights_path)
            self.models_loaded = True
            logger.info(f"Vision model '{self.model_name}' simulated loaded successfully.")
        except Exception as e:
            logger.error(f"Failed to load vision model {self.model_name}: {e}")
            self.models_loaded = False

    async def analyze_video(self, video_path: str) -> Dict[str, Any]:
        """
        Analyzes a video file for objects, events, or anomalies.
        This is a simplified mock implementation.
        """
        if not self.vision_enabled:
            return {"status": "failed", "reason": "Vision engine is disabled."}
        if not self.models_loaded:
            return {"status": "failed", "reason": "Vision models not loaded."}

        video_file = Path(video_path)
        if not video_file.exists():
            logger.error(f"Video file not found: {video_path}")
            return {"status": "failed", "reason": "Video file not found."}

        logger.info(f"Starting analysis of video: {video_path}")
        start_time = datetime.now()

        try:
            # Simulate video processing
            cap = cv2.VideoCapture(str(video_file))
            if not cap.isOpened():
                logger.error(f"Could not open video file: {video_path}")
                return {"status": "failed", "reason": "Could not open video file."}

            frame_count = 0
            detected_objects = {}
            anomalies_detected = False

            while True:
                ret, frame = cap.read()
                if not ret:
                    break
                frame_count += 1

                # Simulate object detection (e.g., person, vehicle, suspicious object)
                if frame_count % 10 == 0: # Process every 10th frame
                    # Dummy detection logic
                    if "security_footage" in video_path:
                        if frame_count > 50 and frame_count < 100:
                            detected_objects["person"] = detected_objects.get("person", 0) + 1
                        if frame_count > 150 and frame_count < 200:
                            detected_objects["unauthorized_access"] = detected_objects.get("unauthorized_access", 0) + 1
                            anomalies_detected = True
                    elif "network_traffic" in video_path:
                        if frame_count % 30 == 0:
                            detected_objects["data_packet"] = detected_objects.get("data_packet", 0) + 1
                        if frame_count > 100 and frame_count < 120:
                            anomalies_detected = True # Simulate a network anomaly

            cap.release()
            end_time = datetime.now()
            analysis_duration = (end_time - start_time).total_seconds()

            result = {
                "status": "completed",
                "video_path": video_path,
                "total_frames": frame_count,
                "analysis_duration_seconds": analysis_duration,
                "detected_objects": detected_objects,
                "anomalies_detected": anomalies_detected,
                "summary": f"Analyzed {frame_count} frames. Detected {sum(detected_objects.values())} objects. Anomalies: {anomalies_detected}.",
                "timestamp": datetime.now().isoformat()
            }
            logger.info(f"Video analysis completed for {video_path}. Status: {result['status']}")
            return result

        except Exception as e:
            logger.error(f"Error during video analysis of {video_path}: {e}")
            return {"status": "failed", "reason": str(e)}

    def get_vision_stats(self) -> Dict[str, Any]:
        """Returns statistics about the vision engine's status."""
        return {
            "enabled": self.vision_enabled,
            "models_loaded": self.models_loaded,
            "model_name": self.model_name,
            "cache_size": f"{self.cache_size_mb}MB (mock)", # Placeholder for actual cache usage
            "recording_active": self.recording_active,
            "last_analysis_timestamp": datetime.now().isoformat() # Placeholder
        }
