requirements.txt
loguru>=0.7.2
PyYAML>=6.0.1
transformers>=4.30.0
torch==2.3.1
chromadb>=0.4.14
sentence-transformers>=2.2.2
aiohttp==3.9.5
beautifulsoup4==4.12.3
requests==2.32.3
gradio>=4.0.0
httpx>=0.25.0
opencv-python>=4.8.1.78 # For Vision Engine
# Optional dependencies for full functionality:
# For Voice Interface:
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyAudio>=0.2.11 # Often requires system-level installation (e.g., portaudio-dev on Linux)
# For Vision Engine (advanced models, screen recording):
# mss==9.0.1
# Pillow==10.3.0
# pytesseract==0.3.10
# pyautogui==0.9.54
# For Docker/Kubernetes Deployment:
# docker==7.1.0
# kubernetes==30.1.0
# For Neo4j (alternative memory backend):
# neo4j==5.21.0
# For Charting in Admin Dashboard:
plotly>=5.18.0
pandas>=2.1.4
# For mocking requests in tests:
# requests-mock==1.12.1
tiktoken>=0.5.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
